backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.27.0
sqlalchemy==2.0.30
psycopg[binary]==3.1.18
pydantic==2.7.1
pdfplumber==0.11.1
sentence-transformers==2.7.0
# scikit-learn==1.4.2  # skipped due to Python 3.13 compatibility issues
spacy==3.7.4
numpy==1.26.4
pandas==2.2.2
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.0

# Authentication & Security
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
slowapi==0.1.9

# Database Migrations
alembic==1.13.1

# HTTP Client (for email service integration)
httpx==0.25.2
